backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
elasticsearch==8.11.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
python-dotenv==1.0.0
httpx==0.25.2
asyncpg==0.29.0
databases==0.8.0
pytest==7.4.3
pytest-asyncio==0.21.1